This PR * Separates the MessageStream from Marigold to have its own class * adds the MessageStream module to Marigold Package
https://sailthru.atlassian.net/browse/MP-1668
